// src/frontend/src/features/ctr-monitoring/types.ts
// Generated by AI on 03/15/2026

export interface FailedCtr {
  taxPacketGuid: string;
  pquid: number;
  clientId: number;
  year: number;
  quarter: number;
  processStatus: string;
  message: string;
  meta_DateCreated: string;
  reportType: string;
  isReportWizardCTR: boolean;
}

export interface FailedForm {
  taxPacketGuid: string;
  id: string;
  reportName: string;
  reportStatus: string;
  reportDataTransStatus: string;
  pdfGenStatus: string;
  reportDataTransMessage: string;
  pdfGenMessage: string;
  taxAuthCode: string;
  taxCode: string;
  taxFormName: string;
  generationType: string;
  isFailedBecauseOfTransientError: boolean;
}